+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rename Design entity Treatments tab to Statistical Design. [#418](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/418)
 - Move design entity fields from Description to Statistical Design tab. [#417](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/417)
 - Change label and description of design entity name field. [#430](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/430)
+- Move restriction and management fields from plan to design entity. [#419](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/419)
 
 ### Fixed